Separate tests into files for each contract used
